demo/services/api/src/requirements.txt
alembic==1.2.1
aniso8601==8.0.0
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
Click==7.0
decorator==4.4.0
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.9
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==2.6.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.3
marshmallow-sqlalchemy==0.19.0
mccabe==0.6.1
mistune==0.8.4
parso==0.5.1
passlib==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.3
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint-plugin-utils==0.6
pyrsistent==0.15.5
python-dateutil==2.8.0
python-editor==1.0.4
python-json-logger==0.1.11
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
requests==2.23.0
six==1.12.0
SQLAlchemy==1.3.10
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
